[refactor] 장소 검색 LazyPaging 로직 개선 (#34)
* refactor : collect 로직 개선

* style: 코드 포맷 수정 & 린트 체크
